requirements.txt
streamlit==1.32.0
google-generativeai==0.8.5
faiss-cpu==1.12.0
PyPDF2==3.0.1
plotly==5.5.0
pandas==2.2.0
python-dotenv==1.0.0
numpy==1.26.0
pytest==8.1.1
pytest-cov==5.0.0
black==24.3.0
flake8==7.0.0
mypy==1.9.0
pre-commit==3.6.2
isort==5.13.2